# LMS-BACKEND/learning-management-service/src/main/java/com/lm/app/model/User.java
package com.lm.app.model;

import java.io.Serializable;
import javax.persistence.*;


@Entity
@Table(name="user")
@NamedQuery(name="User.findAll", query="SELECT a FROM User a")
public class User implements Serializable {
	private static final long serialVersionUID = 1L;

	@Id
	@GeneratedValue(strategy=GenerationType.IDENTITY)
	@Column(name="user_id", unique=true, nullable=false)
	private Integer userId;

	@Column(length=250)
	private String password;
	
	@Column(length=250)
	private String email;

	@Column(length=250)
	private String role;

	@Column(name="user_name", length=250)
	private String userName;

	public User() {
	}

	public Integer getUserId() {
		return this.userId;
	}

	public void setUserId(Integer userId) {
		this.userId = userId;
	}

	public String getPassword() {
		return this.password;
	}

	public void setPassword(String password) {
		this.password = password;
	}

	public String getRole() {
		return this.role;
	}

	public void setRole(String role) {
		this.role = role;
	}

	public String getUserName() {
		return this.userName;
	}

	public void setUserName(String userName) {
		this.userName = userName;
	}

	public String getEmail() {
		return email;
	}

	public void setEmail(String email) {
		this.email = email;
	}
}